+++ src/lib/server/container/types.ts
@@ -13,6 +13,7 @@ import type { IMembrosRepository } from "@/lib/server/db/interfaces/membros-repo
 import type { ICargosRepository } from "@/lib/server/db/interfaces/cargos-repository.interface";
 import type { ICurriculosRepository } from "@/lib/server/db/interfaces/curriculos-repository.interface";
 import type { ICalendarioRepository } from "@/lib/server/db/interfaces/calendario-repository.interface";
+import type { IVagasRepository } from "@/lib/server/db/interfaces/vagas-repository.interface";
 import type { IDisciplinaSemestreRepository } from "@/lib/server/db/interfaces/disciplina-semestre-repository.interface";
 import type { IDisciplinaRepository } from "@/lib/server/db/interfaces/disciplina-repository.interface";
 import type { IDisciplinaConcluidaRepository } from "@/lib/server/db/interfaces/disciplina-concluida-repository.interface";
@@ -40,6 +41,7 @@ export type Container = {
   cargosRepository: ICargosRepository;
   curriculosRepository: ICurriculosRepository;
   calendarioRepository: ICalendarioRepository;
+  vagasRepository: IVagasRepository;
   disciplinaSemestreRepository: IDisciplinaSemestreRepository;
   disciplinaRepository: IDisciplinaRepository;
   disciplinaConcluidaRepository: IDisciplinaConcluidaRepository;